red_panda/requirements.txt
numpy~=1.26.2
torch~=2.1.1+cu121
torchvision~=0.16.1+cu121
matplotlib~=3.8.2
scikit-image~=0.22.0
Pillow~=10.1.0
scikit-learn~=1.3.2
seaborn~=0.13.0
pandas~=2.1.3